tweaks, begin new download.js dev
