+++ dvh/download_new.js
@@ -1,0 +1,20 @@
+// Built from Bokeh Export CSV Example: 2017.04.27
+// https://github.com/bokeh/bokeh/tree/master/examples/app/export_csv
+
+var filename = 'data_result.csv';
+var blob = new Blob([source.data['text'][0]], { type: 'text/csv;charset=utf-8;' });
+
+//addresses IE
+if (navigator.msSaveBlob) {
+    navigator.msSaveBlob(blob, filename);
+}
+
+else {
+    var link = document.createElement("a");
+    link = document.createElement('a')
+    link.href = URL.createObjectURL(blob);
+    link.download = filename
+    link.target = "_blank";
+    link.style.visibility = 'hidden';
+    link.dispatchEvent(new MouseEvent('click'))
+}
