+++ dvh/main.py
@@ -53,7 +53,7 @@ data_tables = DataTables(sources)
 planning_data = PlanningData(custom_title, data_tables)
 roi_viewer = ROI_Viewer(sources, custom_title)
 mlc_analyzer = MLC_Analyzer(sources, custom_title, data_tables)
-time_series = TimeSeries(sources, categories.range, custom_title, data_tables)
+time_series = TimeSeries(sources, categories.range, custom_title)
 correlation = Correlation(sources, categories, custom_title)
 regression = Regression(sources, time_series, correlation, categories.multi_var_reg_var_names, custom_title, data_tables)
 correlation.add_regression_link(regression)
